+++ examples/data-sources/scalr_vcs_provider/data-source.tf
@@ -1,9 +1,7 @@
 data "scalr_vcs_provider" "example1" {
-  id         = "vcs-xxxxxxxxxx"
-  account_id = "acc-xxxxxxxxxx"
+  id = "vcs-xxxxxxxxxx"
 }
 
 data "scalr_vcs_provider" "example2" {
-  name       = "example"
-  account_id = "acc-xxxxxxxxxx"
+  name = "example"
 }